+++ src/commands/browseWebsite.ts
@@ -3,18 +3,16 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { type IActionContext } from '@microsoft/vscode-azext-utils';
-import { functionFilter } from '../constants';
-import { ext } from '../extensionVariables';
+import { nonNullValueAndProp, type IActionContext } from '@microsoft/vscode-azext-utils';
 import { type SlotTreeItem } from '../tree/SlotTreeItem';
+import { type ContainerTreeItem } from '../tree/containerizedFunctionApp/ContainerTreeItem';
 import { openUrl } from '../utils/openUrl';
+import { pickAppResource } from '../utils/pickAppResource';
 
-export async function browseWebsite(context: IActionContext, node?: SlotTreeItem): Promise<void> {
+export async function browseWebsite(context: IActionContext, node?: SlotTreeItem | ContainerTreeItem): Promise<void> {
     if (!node) {
-        node = await ext.rgApi.pickAppResource<SlotTreeItem>(context, {
-            filter: functionFilter,
-        });
+        node = await pickAppResource(context);
     }
 
-    await openUrl(node.site.defaultHostUrl);
+    await openUrl(nonNullValueAndProp(node.site, 'defaultHostUrl'));
 }