+++ src/commands/createNewProject/ProjectCreateStep/DotnetProjectCreateStep.ts
@@ -3,10 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AzExtFsExtra, DialogResponses, type IActionContext } from '@microsoft/vscode-azext-utils';
+import { AzExtFsExtra, DialogResponses, nonNullValueAndProp, type IActionContext } from '@microsoft/vscode-azext-utils';
 import * as path from 'path';
 import { getMajorVersion, type FuncVersion } from '../../../FuncVersion';
 import { ConnectionKey, ProjectLanguage, gitignoreFileName, hostFileName, localSettingsFileName } from '../../../constants';
+import { ext } from '../../../extensionVariables';
 import { MismatchBehavior, setLocalAppSetting } from '../../../funcConfig/local.settings';
 import { localize } from "../../../localize";
 import { executeDotnetTemplateCommand, validateDotnetInstalled } from '../../../templates/dotnet/executeDotnetTemplateCommand';
@@ -31,10 +32,20 @@ export class DotnetProjectCreateStep extends ProjectCreateStepBase {
 
         const projectName: string = path.basename(context.projectPath);
         const projName: string = projectName + language === ProjectLanguage.FSharp ? '.fsproj' : '.csproj';
-        await this.confirmOverwriteExisting(context, projName);
+
+        const workerRuntime = nonNullProp(context, 'workerRuntime');
+        // For containerized function apps we need to call func init before intialization as we want the .csproj file to be overwritten with the correct version
+        // currentely the version created by func init is behind the template version
+        if (context.containerizedProject) {
+            const runtime = context.workerRuntime?.capabilities.includes('isolated') ? 'dotnet-isolated' : 'dotnet';
+            // targetFramework is only supported for dotnet-isolated projects
+            const targetFramework = runtime === 'dotnet' ? '' : "--target-framework " + nonNullValueAndProp(context.workerRuntime, 'targetFramework');
+            await cpUtils.executeCommand(ext.outputChannel, context.projectPath, "func", "init", "--worker-runtime", runtime, targetFramework, "--docker");
+        } else {
+            await this.confirmOverwriteExisting(context, projName);
+        }
 
         const majorVersion: string = getMajorVersion(version);
-        const workerRuntime = nonNullProp(context, 'workerRuntime');
         let identity: string = workerRuntime.projectTemplateId.csharp;
         if (language === ProjectLanguage.FSharp) {
             identity = identity.replace('CSharp', 'FSharp'); // they don't have FSharp in the feed yet