+++ src/tree/ResolvedFunctionAppResource.ts
@@ -8,7 +8,6 @@ import { DeleteLastServicePlanStep, DeleteSiteStep, DeploymentTreeItem, Deployme
 import { AppSettingTreeItem, AppSettingsTreeItem } from "@microsoft/vscode-azext-azureappsettings";
 import { AzureWizard, DeleteConfirmationStep, nonNullValue, type AzExtTreeItem, type IActionContext, type ISubscriptionContext, type TreeItemIconPath } from "@microsoft/vscode-azext-utils";
 import { type ResolvedAppResourceBase } from "@microsoft/vscode-azext-utils/hostapi";
-import { type ViewPropertiesModel } from "@microsoft/vscode-azureresources-api";
 import { latestGAVersion, tryParseFuncVersion, type FuncVersion } from "../FuncVersion";
 import { runFromPackageKey } from "../constants";
 import { ext } from "../extensionVariables";
@@ -17,7 +16,8 @@ import { localize } from "../localize";
 import { createActivityContext } from "../utils/activityUtils";
 import { envUtils } from "../utils/envUtils";
 import { treeUtils } from "../utils/treeUtils";
-import { type ApplicationSettings, type FuncHostRequest } from "./IProjectTreeItem";
+import { type ApplicationSettings } from "./IProjectTreeItem";
+import { ResolvedFunctionAppBase } from "./ResolvedFunctionAppBase";
 import { type SlotTreeItem } from "./SlotTreeItem";
 import { SlotsTreeItem } from "./SlotsTreeItem";
 import { ProjectResource, ProjectSource, matchesAnyPart } from "./projectContextValues";
@@ -27,7 +27,7 @@ export function isResolvedFunctionApp(ti: unknown): ti is ResolvedAppResourceBas
     return (ti as unknown as ResolvedFunctionAppResource).instance === ResolvedFunctionAppResource.instance;
 }
 
-export class ResolvedFunctionAppResource implements ResolvedAppResourceBase {
+export class ResolvedFunctionAppResource extends ResolvedFunctionAppBase implements ResolvedAppResourceBase {
     public site: ParsedSite;
     public data: Site;
 
@@ -61,7 +61,7 @@ export class ResolvedFunctionAppResource implements ResolvedAppResourceBase {
     commandArgs?: unknown[] | undefined;
 
     public constructor(subscription: ISubscriptionContext, site: Site) {
-        this.site = new ParsedSite(site, subscription);
+        super(new ParsedSite(site, subscription))
         this.data = this.site.rawSite;
         this._subscription = subscription;
         this.contextValuesToAdd = [this.site.isSlot ? ResolvedFunctionAppResource.slotContextValue : ResolvedFunctionAppResource.productionContextValue];
@@ -86,24 +86,12 @@ export class ResolvedFunctionAppResource implements ResolvedAppResourceBase {
         return resource;
     }
 
-    public get name(): string {
-        return this.label;
-    }
-
     public get label(): string {
         return this.site.slotName ?? this.site.fullName;
     }
 
-    public get id(): string {
-        return this.site.id;
-    }
-
     public get logStreamLabel(): string {
         return this.site.fullName;
-    }
-
-    public async getHostRequest(): Promise<FuncHostRequest> {
-        return { url: this.site.defaultHostUrl };
     }
 
     public get description(): string | undefined {
@@ -115,19 +103,8 @@ export class ResolvedFunctionAppResource implements ResolvedAppResourceBase {
         return treeUtils.getIconPath(proxyTree.contextValue);
     }
 
-    public get viewProperties(): ViewPropertiesModel {
-        return {
-            data: this.site,
-            label: this.name,
-        }
-    }
-
     private get _state(): string | undefined {
         return this.site.rawSite.state;
-    }
-
-    public hasMoreChildrenImpl(): boolean {
-        return false;
     }
 
     /**
@@ -326,7 +303,7 @@ export class ResolvedFunctionAppResource implements ResolvedAppResourceBase {
     }
 }
 
-function matchContextValue(expectedContextValue: RegExp | string, matches: (string | RegExp)[]): boolean {
+export function matchContextValue(expectedContextValue: RegExp | string, matches: (string | RegExp)[]): boolean {
     if (expectedContextValue instanceof RegExp) {
         return matches.some((match) => {
             if (match instanceof RegExp) {